refactor: rename package model -> models & converter -> responseconverters
