+++ data/src/main/java/com/nemo/data/models/GithubProject.kt
@@ -1,4 +1,4 @@
-package com.nemo.data.model
+package com.nemo.data.models
 
 import java.time.ZonedDateTime
